parser: Fix: Not all numbers with leading zeroes are octal
Prior to this change, all numbers with a leading zero would be identified as
octal, which is incorrect in many ways, and in particular makes floating point
number values starting with zero, e.g. "0.1" to not be recognized as floats,
but as octals instead, which then caused a conversion error. This patch adds
a stricter check for octal numbers, and solves the issue.