Avoid writing too many comma separators for dictionaries
Some comma separators are not really needed, taking into account that
whitespace can also act as a separator. Normally dictionaries would be written
this way:

  # Normal mode
  person {
    name: "Peter",
    age: 32,
  },

  # Compact mode
  person{name:"Peter",age:32,},

This patch avoids the trailing commas in dictionaries: normal mode uses no
commas altogether, because the trailing newline character already acts as item
separator, and in compact mode it avoids the trailing comma in the last item.
With this patch applied, the output is:

  # Normal mode
  person {
    name: "Peter"
    age: 32
  }

  # Compact mode
  person{name:"Peter",age:32}