Ensure all numeric constraints in parse_number loop
Instead of leaving some of the checks to be done later on regarding number
formats, ensure in the loop that constraints are satisfied for octals. For
the rest of numeric formats, those were already being enforced, so checking
for octals allows to remove the checks before conversion, and use assert()
instead just for debug builds.