Define missing functions to create hipack_value_t objects
Add a missing expansion of the HIPACK_TYPES macro which calls the
HIPACK_DEFINE_MAKE_VALUE one to define the functions that may be used
to create hipack_value_t objects.

While at it, fix the HIPACK_DEFINE_MAKE_VALUE macro to use a designated
initializer that ensures the correct union member is used.